--- src/components/Dashboard.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Dashboard = () => (
-  <div className="block-group">
-    <div className="block block-content">
-      <h1>Dashboard</h1>
-      <p>Now you can start building your custom React app</p>
-    </div>
-  </div>
-);
-
-export default Dashboard;
