+++ src/components/Header/Header.jsx
@@ -2,11 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const Header = ({ appName }) => (
-  <header>
-    <div className="block-group">
-      <div className="block block-logo">
-        <h1>{ appName }</h1>
-      </div>
+  <header className="header">
+    <div className="header--logo-wrapper">
+      <h1 className="header--logo-title">{ appName }</h1>
     </div>
   </header>
 );
