+++ src/less/body/base.less
@@ -52,14 +52,21 @@ table {
 
 p {
   margin: 1em 0;
-  // #font.type();
 }
 
 h1 {
-	font-size: 2.6em;
+	font-size: 2.2rem;
 }
 
 h2 {
   font-size: 1.6em;
   margin: 1em 0;
+}
+
+main {
+	#display.grid();
+}
+
+#webpack-dev-server-client-overlay {
+ background: #000 !important; 
 }
