+++ src/less/components/menu.less
@@ -1,0 +1,21 @@
+.menu {
+  background: @grey-lt;
+  height: 82vh;
+  padding-top: 4px;
+  #grid.column(1, span 2);
+  
+  &--list-item-link {
+    display: block;
+    width: 90%;
+    padding: .5rem 0;
+    text-decoration: none;
+    padding-left: 1rem;
+    #border.solid(bottom, 2px, @white);
+    #display.colors(@white, darken(@blue, 40%));
+
+    &:hover,
+    &---active {
+      background-color: lighten(@blue, 5%);
+    }
+  }
+}
