--- src/less/components/nav.less
@@ -1,20 +1,0 @@
-nav {
-  background: @grey-lt;
-  padding: 1em 0;
-  
-  ul {
-    #display.clearfix();
-    
-    li {
-
-      a {
-        #button.link(@white, @blue);
-        
-        &:hover,
-        &.active {
-          background: darken(@blue, 40%);
-        }
-      }
-    }
-  }
-}
