+++ src/less/imports.less
@@ -9,11 +9,9 @@
 // All default styling, browser reset
 @import './body/base';
 
-@import './libs/grid';
-
 // Import componenets. Eg: Header, Navigation, Footer etc
 @import './components/header';
-@import './components/nav';
+@import './components/menu';
 
 @import './body/content';
