--- src/less/libs/grid.less
@@ -1,52 +1,0 @@
-// Pocketgrid : http://arnaudleray.github.io/pocketgrid/docs
-
-.block,
-.block:after,
-.block:before
-.block-group,
-.block-group:before,
-.block-group:after {
-  -webkit-box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  box-sizing: border-box;
-}
-
-.block {
-  float: left;
-  width: 100%;
-  // padding: 2px 6px; // Gutter
-}
-
-.block-group {
-  *zoom: 1;
-  list-style-type: none;
-  padding: 0;
-  // margin: -2px -6px; // Gutter
-  margin: 0 auto;
-  width: 100%;
-
-  @media @desktop-narrow {
-    width: 992px;
-  }
-
-  @media @desktop-wide {
-    width: 1200px;
-  }
-
-  &:before,
-  &:after {
-    display: table;
-    content: "";
-    line-height: 0;
-  }
-
-  &:after {
-    clear: both;
-  }
-
-  > .block-group {
-    clear: none;
-    float: left;
-    margin: 0 !important;
-  }
-}
