+++ src/pages/About.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const About = () => (
+  <div className="content content-about">
+    <h1>About this package:</h1>
+    <p>This package contains a default boilerplate project set up with some nice dev tools.</p>
+  </div>
+);
+
+export default About;
