+++ src/pages/Dashboard.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Dashboard = () => (
+  <div className="content content-dashboard">
+    <h1>Dashboard</h1>
+    <h2>Thank you for your download.</h2>
+    <p>
+      This is my template kit for starting new apps.
+      I have some custom architecture which I find works for me.
+    </p>
+  </div>
+);
+
+export default Dashboard;
